+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Optional support for sentry error collector.
+
 ## [1.0.2] - 2020-06-18
 
 ### Fixed